requirements.txt
# ── Orchestration ─────────────────────────────────────────────────────────────
crewai>=0.28.0

# ── LangChain (used by CrewAI under the hood) ─────────────────────────────────
langchain>=0.3.0
langchain-openai>=0.2.0

# ── LangSmith (optional observability) ────────────────────────────────────────
langsmith>=0.1.0

# ── LLM provider ──────────────────────────────────────────────────────────────
openai>=1.0.0

# ── Brokerage ─────────────────────────────────────────────────────────────────
alpaca-py>=0.20.0

# ── Market data ───────────────────────────────────────────────────────────────
alpha-vantage>=2.3.3
yfinance>=0.2.40             # free fallback when Alpha Vantage rate limit is hit

# ── News ──────────────────────────────────────────────────────────────────────
newsapi-python>=0.2.7

# ── Technical analysis (pure Python, no C compilation required) ───────────────
ta>=0.11.0

# ── Scraping ──────────────────────────────────────────────────────────────────
playwright>=1.44.0
beautifulsoup4>=4.12.0
lxml>=5.2.0

# ── Data ──────────────────────────────────────────────────────────────────────
pandas>=2.2.0
numpy>=1.26.0,<2.0.0        # numpy 2.x breaks some TA library binary wheels

# ── Schemas / validation ──────────────────────────────────────────────────────
pydantic>=2.7.0
pydantic-settings>=2.3.0

# ── Redis ─────────────────────────────────────────────────────────────────────
redis>=5.0.0

# ── Scheduling ────────────────────────────────────────────────────────────────
apscheduler>=3.10.0

# ── Sentiment ─────────────────────────────────────────────────────────────────
vaderSentiment>=3.3.2

# ── Networking ────────────────────────────────────────────────────────────────
httpx>=0.27.0
requests>=2.32.0

# ── Config ────────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
pytz>=2024.1

# ── Observability ─────────────────────────────────────────────────────────────
structlog>=24.1.0
